Feature/batch update (#23)
* Refactor RepoUpdateService to handle multiple repositories with same URL

Refactored the `RepoUpdateService` to accommodate multiple repositories with the same URL. This includes replacing `Repository` with a string URL for the `priorityQueue` and `repositoryQueue`, and introducing `updateRepository` method to update individual repositories. We also changed the error handling to log any errors and updated the update logic to operate on repositories instead of URL. This change was necessary as it allows for better updating of multiple repositories that share a same URL.

* Add updateIdentifier method to RawRepository and correct formatting

Added a new method, updateIdentifier, to the RawRepository which allows for the repository identifier to be updated while maintaining the same url. This addresses cases where the identifier might need to be updated but the URL remains the same. Also, made several minor formatting corrections including consistent spaces after commas for better readability.

* Add equalsRepository method in Identifier class

The new method, equalsRepository, has been added to the Identifier class. This method compares the platform, user, and repository of two Identifier objects and returns true if they are equal, and false otherwise. This was necessary to improve the accuracy and efficiency of comparison between different Identifier objects.

* Add logging to Data class

Logger and logging messages were added to the Data class. This was done to improve tracking repository update checks and failures in the application by providing meaningful debug logging. These logs provide insight on which repository was checked for updates and what caused an update to fail.

* Change return types of 'byUrl' and 'leastUpdated' methods

'byUrl' method's return type has been modified from Optional<Repository> to List<Repository> to provide all repositories with a specified URL instead of only the first one. 'leastUpdated' method now returns a list of URLs (List<String>) instead of Repository objects, minimizing data transmission need and simplifying the method's use. This change mainly focuses on improving method usability and efficiency.

* Add view for repository updates in PostgreSQL

Added a SQL script (patch_5.sql) which creates a PostgreSQL view that aggregates update information for each repository. This change was necessary for a feature that allows monitoring repository update progress on the application. The version file was updated to reflect the changes in the database schema.