Add new repository_updates view and refactor identifier column in repository table
Introduced a new view, repository_updates, that groups repository data by URL and finds the minimum checked timestamp for each group. This view will provide a better overview of the latest updates for each repository.

Simultaneously, the identifier column in the repository table has been refactored. This change handles the cases when the repository path includes slashes, the identifier will not be wrongly segmented. This aims to better parse the repository identifier when it includes a path with lashes.